+++ packages/freezed/test/common.dart
@@ -1,5 +1,3 @@
-// @dart=2.9
-
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/diagnostic/diagnostic.dart';
 import 'package:analyzer/error/error.dart';