+++ packages/freezed/test/json_test.dart
@@ -1,5 +1,3 @@
-// @dart=2.9
-
 // ignore_for_file: prefer_const_constructors, omit_local_variable_types
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:build_test/build_test.dart';
@@ -341,7 +339,7 @@ Future<void> main() async {
     });
 
     test('tear-off', () {
-      Generic<int> Function(Map<String, Object>) fromJson = $Generic.fromJson;
+      Generic<int> Function(Map<String, Object?>) fromJson = $Generic.fromJson;
 
       expect(
         fromJson(<String, dynamic>{'a': 42}),